http-ignition/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
proxmoxer==2.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
pythonping==1.1.4
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
scapy==2.6.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.2
typer==0.12.5
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==75.3.0.20241105
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1